requirements.txt
langchain==0.1.14
langchain-community==0.0.30
openai==1.30.1
streamlit==1.35.0
python-dotenv==1.0.1
